refactor(EditorSidebar): split into panels/sections and a hook (#196)
- Extract background option lists/types to editorSidebar/backgroundOptions.ts
- Move background local state, effects and handlers into
  editorSidebar/useEditorSidebarBackground.ts
- Extract SidebarHeader, BackgroundPanel, ImageBackgroundTab, ToolsPanel and
  AddTextSection sub-components
